storage: add support for v3 policy and IAM conditions
Adds GetWithVersion for future IAM conditions support.

After this is merged, a new version of the storage module should be
tagged, and the root cloud.google.com/go module updated to depend on
that latest storage module.

This means we can add the GetWithVersion method to the interface without
breaking compatibility.

See CL 44730 for the followup change.

Change-Id: I83ffb99e0581781d99ad54ead50c2a5e0682cc83
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/52390
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Cody Oss <codyoss@google.com>