spanner/spannertest: disable logging in session RPCs
These have become quite noisy due to some recent changes in the Spanner
client, and these aren't generally useful to see for most users of
spannertest. They've been left in place, commented out, so it is easier
to re-enable for future maintainers of the Spanner client or spannertest
who want to diagnose issues to do with sessions.

Change-Id: Ia01541726edb86b3cb4c89c2878ea58e19e4f74e
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/48310
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Knut Olav Løite <koloite@gmail.com>