spanner: Make logger configurable
The Spanner client can now be configured to use a specific
logger instead of the standard logger. If no specific logger
is configured, the client will use the standard logger. This
enables test cases to use a custom logger to filter expected
errors, keeping the test logs as clean as possible.

Fixes #1678.

Change-Id: I3be3d0a70c6941b5cf1eb514c67764bed02644e6
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/48731
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>
Reviewed-by: Shanika Kuruppu <skuruppu@google.com>