spanner: Use Go 1.13 error-unwrapping
Use Go 1.13 error-unwrapping and the equivalent xerrors features
for builds on versions < 1.13. This makes it possible to use
wrapped errors with the Spanner client.

This change also deprecates the gRPC code in the Spanner error struct.
All functions that need the gRPC code will extract it from the wrapped
error instead of reading this field.

Fixes #1223 and #1608.
Updates #1310.

Change-Id: Iea914adb5ca78af5e78cc948d8b1180eb3d647d2
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/48730
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hengfeng Li <hengfeng@google.com>