firestore: migrate admin v1 API to microgen
The generated client is included in this CL because it contains a breaking change.
Manual changes to the mock_test.go were necessary to account for the
breaking change and it is easier to ship the change as a unit.

This client is still in the alpha stage, so the breaking change is
acceptable. The actual breakage is in several RPCs that previously
returned a longrunningpb.Operation directly and are now returning a
response-specific Operation type with helpers.

Change-Id: Ia016f8a0b9e797aac228322594f0b82cc394a448
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/49534
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chris Cotter <cjcotter@google.com>