internal/gapicgen: genmgr update go.mod frequently
Currently if genmgr runs right after genbot, the go module proxies may not have
picked up the latest version of genproto, resulting in go.mod not getting
latest genproto BUT the CL is updated with reviewers. So, the only way to get
the CL to have updated go.mod is to remove reviewers, run genmgr again, which
will cause it get past its !hasReviewers check and update go.mod and add
reviewers again.

This CL changes that behavior such that it always tries to update go.mod
regardless of if there are reviewers or not. Separately it will add reviewers if
there are none.

Also some formatting fixes and changes some exit 1s to exit 0s.

Change-Id: I879feea18f77090a0fcddfea16147eaf694ffd44
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/49810
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>