gapicgen: simplify setup and documentation.
The cookie name used for gerrit authentication is standard, and so this
change sets the default to that value ("o") and removes references to it
in documentation.

This PR also attempts to clarify what the relevant auth token from
Gerrit code review looks like in the documentation, in hopes that future
users don't get tripped on this as much.

Change-Id: Ied4f780f1246b4269ba3f9c48e6958ec57593687
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/50611
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Alex Hong <hongalex@google.com>