spanner/spansql: preserve filename, token position information during parse
This is useful for tools built on top of spansql to be able to construct
cross-references and similar. It is also a necessary condition for tools
such as a formatter.

This necessitates some backward incompatible changes to spansql, but
they are simple to track. In particular, ParseDDL takes a filename
argument, and a bunch of the AST types are now pointers instead of
values. This has no impact on users of spannertest, only on direct users
of the spansql package.

Change-Id: Ic21e83879cdb1da8ac04692a5198e311383ca7a2
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/50850
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Knut Olav Løite <koloite@gmail.com>