spanner: fix race condition in batch read-only tx
BatchReadOnlyTransaction contained a possible race condition as the
same read or query request would be re-used for the different partitions
of a transaction. That could cause one of the partitions to execute the
wrong query, as it could be assigned the partition token of a different
partition if the partitions were being used in parallel by the user
application.

Fixes #1895.

Change-Id: Ie79c58d843ca59d1259f8fc7a78b00d9e7fa1d40
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/54190
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hengfeng Li <hengfeng@google.com>