storage: add Exists examples for BucketHandle, ObjectHandle
While there isn't a top level method "Exists" for either
of BucketHandle or ObjectHandle, the use of Attrs seems to
require one to already be familiar with these libraries.

In the related issue #1773, a customer tried to create a Reader
just to check if an object existed. They performed an expensive
call instead of just the cheap metadata check.
This CL ensures that such guidance for checking if buckets and objects
exist; these examples will be indexed by search engines and next
time they won't have to go digging through the docs and experimenting
with expensive APIs.

Fixes #1773

Change-Id: I5e06da7304699db3fa5591298b2ac98551222e54
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/54630
Reviewed-by: Cody Oss <codyoss@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>