spanner: retry CreateDatabase on retryable codes
The CreateDatabase RPC should be retried if it fails with a transient
error code, such as UNAVAILABLE or DEADLINE_EXCEEDED. The actual RPC
should only be retried if the initial RPC did not already start an
operation for creating a new database. In case the initial call
succeeded, but the response was lost, the retry logic will return a
reference to the existing long-running operation.

Updates #1964
Fixes #1922 and #1921

Change-Id: I2871b4ade440bd371ba717ca324f6542a6321a59
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/55570
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hengfeng Li <hengfeng@google.com>